Move aws SDK setup to testkit
